# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: RentSaathi - Indian Landlord Rent Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /agreements:
    get:
      operationId: listAgreements
      tags: [agreements]
      summary: List all rental agreements
      responses:
        "200":
          description: Agreements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Agreement"
    post:
      operationId: createAgreement
      tags: [agreements]
      summary: Create a rental agreement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgreementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agreement"
  /agreements/{id}:
    get:
      operationId: getAgreement
      tags: [agreements]
      summary: Get an agreement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Agreement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agreement"
    patch:
      operationId: updateAgreement
      tags: [agreements]
      summary: Update an agreement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgreementUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agreement"
    delete:
      operationId: deleteAgreement
      tags: [agreements]
      summary: Delete an agreement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /maintenance:
    get:
      operationId: listMaintenanceRequests
      tags: [maintenance]
      summary: List all maintenance requests
      responses:
        "200":
          description: Maintenance requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MaintenanceRequest"
    post:
      operationId: createMaintenanceRequest
      tags: [maintenance]
      summary: Raise a maintenance request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaintenanceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceRequest"
  /maintenance/{id}:
    get:
      operationId: getMaintenanceRequest
      tags: [maintenance]
      summary: Get a maintenance request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Maintenance request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceRequest"
    patch:
      operationId: updateMaintenanceRequest
      tags: [maintenance]
      summary: Update a maintenance request status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaintenanceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceRequest"

  # ── Expenses ───────────────────────────────────────────────────────────────
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List all expenses
      responses:
        "200":
          description: Expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Log an expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{id}:
    patch:
      operationId: updateExpense
      tags: [expenses]
      summary: Update an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Agreement:
      type: object
      required:
        - id
        - tenantId
        - unitId
        - type
        - startDate
        - endDate
        - rentAmount
        - depositAmount
        - status
        - createdAt
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        tenantName:
          type: ["string", "null"]
        unitId:
          type: integer
        unitNumber:
          type: ["string", "null"]
        propertyName:
          type: ["string", "null"]
        type:
          type: string
          enum: [residential, shop, pg, room]
        startDate:
          type: string
        endDate:
          type: string
        rentAmount:
          type: number
        depositAmount:
          type: number
        noticePeriodDays:
          type: ["integer", "null"]
        terms:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, expired, terminated, draft]
        daysToExpiry:
          type: ["integer", "null"]
        createdAt:
          type: string
    AgreementInput:
      type: object
      required:
        - tenantId
        - unitId
        - type
        - startDate
        - endDate
        - rentAmount
        - depositAmount
      properties:
        tenantId:
          type: integer
        unitId:
          type: integer
        type:
          type: string
          enum: [residential, shop, pg, room]
        startDate:
          type: string
        endDate:
          type: string
        rentAmount:
          type: number
        depositAmount:
          type: number
        noticePeriodDays:
          type: integer
        terms:
          type: string
    AgreementUpdate:
      type: object
      properties:
        startDate:
          type: string
        endDate:
          type: string
        rentAmount:
          type: number
        depositAmount:
          type: number
        noticePeriodDays:
          type: integer
        terms:
          type: string
        status:
          type: string
          enum: [active, expired, terminated, draft]

    # ── Maintenance schemas ───────────────────────────────────────────────────
    MaintenanceRequest:
      type: object
      required:
        - id
        - unitId
        - tenantId
        - category
        - description
        - status
        - priority
        - createdAt
      properties:
        id:
          type: integer
        unitId:
          type: integer
        unitNumber:
          type: ["string", "null"]
        propertyName:
          type: ["string", "null"]
        tenantId:
          type: integer
        tenantName:
          type: ["string", "null"]
        category:
          type: string
          enum: [plumbing, electrical, carpentry, painting, cleaning, pest_control, other]
        description:
          type: string
        status:
          type: string
          enum: [raised, acknowledged, in_progress, resolved, closed]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        assignedVendorId:
          type: ["integer", "null"]
        assignedVendorName:
          type: ["string", "null"]
        cost:
          type: ["number", "null"]
        deductFromDeposit:
          type: boolean
        resolvedAt:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    MaintenanceInput:
      type: object
      required:
        - unitId
        - tenantId
        - category
        - description
        - priority
      properties:
        unitId:
          type: integer
        tenantId:
          type: integer
        category:
          type: string
          enum: [plumbing, electrical, carpentry, painting, cleaning, pest_control, other]
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high, urgent]
        notes:
          type: string
    MaintenanceUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [raised, acknowledged, in_progress, resolved, closed]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        assignedVendorId:
          type: ["integer", "null"]
        cost:
          type: number
        deductFromDeposit:
          type: boolean
        resolvedAt:
          type: string
        notes:
          type: string

    # ── Expense schemas ────────────────────────────────────────────────────────
    Expense:
      type: object
      required:
        - id
        - propertyId
        - category
        - amount
        - date
        - description
        - createdAt
      properties:
        id:
          type: integer
        propertyId:
          type: integer
        propertyName:
          type: ["string", "null"]
        unitId:
          type: ["integer", "null"]
        unitNumber:
          type: ["string", "null"]
        category:
          type: string
          enum: [repair, painting, property_tax, insurance, maintenance, utilities, other]
        amount:
          type: number
        date:
          type: string
        description:
          type: string
        receiptUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    ExpenseInput:
      type: object
      required:
        - propertyId
        - category
        - amount
        - date
        - description
      properties:
        propertyId:
          type: integer
        unitId:
          type: integer
        category:
          type: string
          enum: [repair, painting, property_tax, insurance, maintenance, utilities, other]
        amount:
          type: number
        date:
          type: string
        description:
          type: string
        receiptUrl:
          type: string
    ExpenseUpdate:
      type: object
      properties:
        category:
          type: string
          enum: [repair, painting, property_tax, insurance, maintenance, utilities, other]
        amount:
          type: number
        date:
          type: string
        description:
          type: string
        receiptUrl:
          type: string

    # ── Vendor schemas ─────────────────────────────────────────────────────────
...