# lib/api-zod/src/generated/types/agreementInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RentSaathi - Indian Landlord Rent Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { AgreementInputType } from './agreementInputType';

export interface AgreementInput {
  tenantId: number;
  unitId: number;
  type: AgreementInputType;
  startDate: string;
  endDate: string;
  rentAmount: number;
  depositAmount: number;
  noticePeriodDays?: number;
  terms?: string;
}