# lib/api-zod/src/generated/types/expenseUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RentSaathi - Indian Landlord Rent Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { ExpenseUpdateCategory } from './expenseUpdateCategory';

export interface ExpenseUpdate {
  category?: ExpenseUpdateCategory;
  amount?: number;
  date?: string;
  description?: string;
  receiptUrl?: string;
}