# lib/api-zod/src/generated/types/maintenanceInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RentSaathi - Indian Landlord Rent Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { MaintenanceInputCategory } from './maintenanceInputCategory';
import type { MaintenanceInputPriority } from './maintenanceInputPriority';

export interface MaintenanceInput {
  unitId: number;
  tenantId: number;
  category: MaintenanceInputCategory;
  description: string;
  priority: MaintenanceInputPriority;
  notes?: string;
}